Fix some issues with thread ordering
We now lock the queue mutex, and then wait for all the accounts
to lock their mutexes before we release the queue mutex. This way
we can guarantee that another transaction that uses the same accounts
doesn't run while we are updating one account. Essentially, it
only allows transactions (either an actual transaction or a check)
to occur if it doesn't interfere with any currently executing
instruction.